requirements.txt
numpy
onnxruntime
opencv-python-headless
Pillow